workers/WorkerGlobalScope_setInterval.htm

<!DOCTYPE html>
<title> WorkerGlobalScope API: setInterval() </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var result = [];
  var worker = new Worker('./support/Timer.js');
  worker.onmessage = t.step_func(function(e) {
    result.push(e.data);
    if (result.length == 3) {
      setTimeout(t.step_func_done(function() {
        assert_array_equals(result, ["hello", "worker", "worker"])
      }), 100);
    }
  });
  worker.postMessage("IntervalHandler");
});
</script>
